fix(gpu-bridge): stop copying the parked escape state into its own closure
Three clang-tidy findings in code this branch just added or reapplied.

run_d3d9_escape took its shared_ptr by value and then copied it again into
the await_host_condition closure; it now moves it, and resolves the thread
and vCPU out of the state first so nothing reads through the moved-from
handle. handle_record_and_call's std::move on the trivially-copyable
io_device_context did nothing. And the park report's parks-per-second divide
narrowed a uint64_t into the duration's double implicitly.

The tidy preset still does not pass on this branch, but every remaining
diagnostic is in code this change does not touch.

+++ src/windows-emulator/wait_storm_diag.hpp
@@ -225,7 +225,7 @@ namespace sogen::wait_storm_diag
         {
             fprintf(stderr,
                     "[WAIT_STORM_DIAG] host_wait_park       parks=%llu (%.0f/s) wake_mean=%.2f ms wake_max=%.2f ms scans/wake=%.1f\n",
-                    static_cast<unsigned long long>(s.host_wait_parks), s.host_wait_parks / dt,
+                    static_cast<unsigned long long>(s.host_wait_parks), static_cast<double>(s.host_wait_parks) / dt,
                     static_cast<double>(s.host_wait_wake_ns) / 1e6 / static_cast<double>(s.host_wait_parks),
                     static_cast<double>(s.host_wait_wake_ns_max) / 1e6,
                     static_cast<double>(s.host_wait_wake_scans) / static_cast<double>(s.host_wait_parks));
